feat: :sparkles: Code fixes due to breaking API changes
Previously, we used advanced typing to adjust the type of path strings without a leading / symbol to flatten RouteConfig objects. However, we determined that this implementation was incorrect - paths should start with / and it resulted in a poor developer experience and inconsistent API. This breaking API fix has resulted in a significant reduction in code and improved performance in the flattenRouteConfig function and FlattenRouteConfing type.