test: :sparkles: split tests to focus on individual features
Previously, all tests were concentrated in a single file, which made it difficult to focus on specific features and increased complexity as the number of tests grew. To address this issue, tests have been split by feature, allowing for more focused testing of individual functionalities. This change improves maintainability and makes the tests easier to read and manage.