fix: make query parameters optional in ExtractRouteData
In the previous version, all properties of the `query` value in `ExtractRouteData` were required by default. To make them optional, a `?` had to be added as a suffix to the query area name.  Now, the `query` value is typed as `Partial<Queries>` when the query area exists, ensuring that query parameters are treated as optional.